+++ src/components/HomePage/HomePage.js
@@ -26,7 +26,7 @@ const HomePage = () => {
                     <HomepageSuggestiongroup />
 
                 </div>
-                <div className="col-span-12 md:col-span-6 px-1 py-2 rounded">
+                <div className="col-span-12 md:col-span-6 px-1 py-2 pb-1 rounded">
 
                     <HomepageCreatepost />
