+++ src/components/HomePage/HomepageAdvertisement.js
@@ -3,12 +3,12 @@ import React from 'react'
 const HomepageAdvertisement = () => {
     return (
         <div>
-            <div className='p-4 pt-0'>
-                <h6>
+            <div className='p-2 pt-2 bg-white rounded m-2'>
+                <h6 className='m-0'>
                     <strong>Advertisement</strong>
                 </h6>
                 <a href='#'>
-                    <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded w-full h-48 pt-3' />
+                    <img src="/user.jpg" alt="" className='object-cover w-full h-48 pt-3' />
                 </a>
             </div>
         </div>
