+++ src/components/HomePage/HomepageBirthday.js
@@ -2,8 +2,8 @@ import React from 'react'
 
 const HomepageBirthday = () => {
     return (
-        <div>
-            <div className='pr-4 pt-2 bg-white m-2 rounded p-4'>
+        <>
+            <div className='bg-white m-2 rounded p-2'>
                 <div className='flex justify-between pb-4'>
                     <h6>
                         <strong>Today Birthdays</strong>
@@ -13,36 +13,32 @@ const HomepageBirthday = () => {
                     </h6>
                 </div>
                 <div>
-                    <a href='#' className='flex items-center mb-6'>
-                        <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded-full w-12 h-12' />
-                        </div>
-                        <div className='ml-4 capitalize'>
-                            <h6><strong>najmul hasan</strong></h6>
+                    <a href='#' className='flex md:flex-col lg:flex-row items-center mb-6'>
+                        <img src="/user.jpg" alt="" className='rounded-full w-12 h-12' />
+
+                        <div className='md:ml-4 text-center capitalize'>
+                            <h6 className='m-0'><strong>najmul hasan</strong></h6>
                             <span>10 minits ago</span>
                         </div>
                     </a>
-                    <a href='#' className='flex items-center mb-6'>
-                        <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded-full w-12 h-12' />
-                        </div>
-                        <div className='ml-4 capitalize'>
-                            <h6><strong>najmul hasan</strong></h6>
+                    <a href='#' className='flex md:flex-col lg:flex-row items-center mb-6'>
+                        <img src="/user.jpg" alt="" className='rounded-full w-12 h-12' />
+                        <div className='md:ml-4 text-center capitalize'>
+                            <h6 className='m-0'><strong>najmul hasan</strong></h6>
                             <span>10 minits ago</span>
                         </div>
                     </a>
-                    <a href='#' className='flex items-center'>
-                        <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded-full w-12 h-12' />
-                        </div>
-                        <div className='ml-4 capitalize'>
-                            <h6><strong>najmul hasan</strong></h6>
+                    <a href='#' className='flex md:flex-col lg:flex-row items-center mb-6'>
+                        <img src="/user.jpg" alt="" className='rounded-full w-12 h-12' />
+
+                        <div className='md:ml-4 text-center capitalize'>
+                            <h6 className='m-0'><strong>najmul hasan</strong></h6>
                             <span>10 minits ago</span>
                         </div>
                     </a>
                 </div>
             </div>
-        </div>
+        </>
     )
 }
