+++ src/components/HomePage/HomepageFollow.js
@@ -10,7 +10,7 @@ const HomepageFollow = () => {
                 <div className='mt-6'>
                     <a href='#' className='flex items-center justify-between mb-4'>
                         <div className='flex items-center'>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded-full w-12 h-12' />
+                            <img src="/user.jpg" alt="" className='rounded-full w-12 h-12' />
 
                             <h6 className='ml-2 capitalize text-base'><strong>najmul hasan</strong></h6>
                         </div>
@@ -18,7 +18,7 @@ const HomepageFollow = () => {
                     </a>
                     <a href='#' className='flex items-center justify-between mb-4'>
                         <div className='flex items-center'>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded-full w-12 h-12' />
+                            <img src="/user.jpg" alt="" className='rounded-full w-12 h-12' />
 
                             <h6 className='ml-2 capitalize text-base'><strong>najmul hasan</strong></h6>
                         </div>
@@ -26,7 +26,7 @@ const HomepageFollow = () => {
                     </a>
                     <a href='#' className='flex items-center justify-between mb-4'>
                         <div className='flex items-center'>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded-full w-12 h-12' />
+                            <img src="/user.jpg" alt="" className='rounded-full w-12 h-12' />
 
                             <h6 className='ml-2 capitalize text-base'><strong>najmul hasan</strong></h6>
                         </div>
@@ -34,7 +34,7 @@ const HomepageFollow = () => {
                     </a>
                     <a href='#' className='flex items-center justify-between mb-4'>
                         <div className='flex items-center'>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded-full w-12 h-12' />
+                            <img src="/user.jpg" alt="" className='rounded-full w-12 h-12' />
 
                             <h6 className='ml-2 capitalize text-base'><strong>najmul hasan</strong></h6>
                         </div>
