+++ src/components/HomePage/HomepageLike.js
@@ -8,39 +8,17 @@ const HomepageLike = () => {
                     <strong>Page You Like</strong>
                 </h6>
                 <div className='mt-4'>
-                    <a href="#" className='flex items-center mb-6'>
-                        <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded-full w-16 h-16' />
-                        </div>
+                    <a href="#" className='sm:block md:flex items-center mb-6'>
+                        <img src="/user.jpg" alt="" className='rounded-full w-16 h-16' />
                         <div className='ml-4 capitalize'>
-                            <h6><strong>Graphic Design</strong></h6>
+                            <h6 className='m-0'><strong>Graphic Design</strong></h6>
                             <span>1215 Members</span>
                         </div>
                     </a>
                     <a href="#" className='flex items-center mb-6'>
-                        <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded-full w-16 h-16' />
-                        </div>
+                        <img src="/user.jpg" alt="" className='rounded-full w-16 h-16' />
                         <div className='ml-4 capitalize'>
-                            <h6><strong>Graphic Design</strong></h6>
-                            <span>1215 Members</span>
-                        </div>
-                    </a>
-                    <a href="#" className='flex items-center mb-6'>
-                        <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded-full w-16 h-16' />
-                        </div>
-                        <div className='ml-4 capitalize'>
-                            <h6><strong>Graphic Design</strong></h6>
-                            <span>1215 Members</span>
-                        </div>
-                    </a>
-                    <a href="#" className='flex items-center mb-6'>
-                        <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded-full w-16 h-16' />
-                        </div>
-                        <div className='ml-4 capitalize'>
-                            <h6><strong>Graphic Design</strong></h6>
+                            <h6 className='m-0'><strong>Graphic Design</strong></h6>
                             <span>1215 Members</span>
                         </div>
                     </a>
