+++ src/components/HomePage/HomepageStory.js
@@ -12,40 +12,40 @@ const HomepageStory = () => {
                         <a href='#' className='underline'>See All</a>
                     </h6>
                 </div>
-                <div className='grid grid-cols-12 mx-auto gap-4'>
+                <div className='grid grid-cols-12 mx-auto gap-2'>
                     <div className='col-span-2 cursor-pointer'>
                         <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded w-72 h-20' />
+                            <img src="/user.jpg" alt="" className='object-contain rounded w-72 h-20' />
                             <span className='capitalize text-sm text-center block py-1'>add story</span>
                         </div>
                     </div>
                     <div className='col-span-2 cursor-pointer'>
                         <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded w-72 h-20' />
+                            <img src="/user.jpg" alt="" className='object-contain rounded w-72 h-20' />
                             <span className='capitalize text-sm text-center block py-1'>najmul hasan</span>
                         </div>
                     </div>
                     <div className='col-span-2 cursor-pointer'>
                         <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded w-72 h-20' />
+                            <img src="/user.jpg" alt="" className='object-contain rounded w-72 h-20' />
                             <span className='capitalize text-sm text-center block py-1'>najmul hasan</span>
                         </div>
                     </div>
                     <div className='col-span-2 cursor-pointer'>
                         <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded w-72 h-20' />
+                            <img src="/user.jpg" alt="" className='object-contain rounded w-72 h-20' />
                             <span className='capitalize text-sm text-center block py-1'>najmul hasan</span>
                         </div>
                     </div>
                     <div className='col-span-2 cursor-pointer'>
                         <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded w-72 h-20' />
+                            <img src="/user.jpg" alt="" className='object-contain rounded w-72 h-20' />
                             <span className='capitalize text-sm text-center block py-1'>najmul hasan</span>
                         </div>
                     </div>
                     <div className='col-span-2 cursor-pointer'>
                         <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded w-72 h-20' />
+                            <img src="/user.jpg" alt="" className='object-contain rounded w-72 h-20' />
                             <span className='capitalize text-sm text-center block py-1'>najmul hasan</span>
                         </div>
                     </div>
