+++ src/components/HomePage/HomepageSuggestiongroup.js
@@ -10,7 +10,7 @@ const HomepageSuggestiongroup = () => {
                 <div className='mt-4'>
                     <a href='#' className='flex items-center mb-6'>
                         <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded w-16 h-16' />
+                            <img src="/user.jpg" alt="" className='rounded w-16 h-16' />
                         </div>
                         <div className='ml-4 capitalize'>
                             <h6><strong>Graphic Design</strong></h6>
@@ -20,7 +20,7 @@ const HomepageSuggestiongroup = () => {
                     </a>
                     <a href='#' className='flex items-center mb-6'>
                         <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded w-16 h-16' />
+                            <img src="/user.jpg" alt="" className='rounded w-16 h-16' />
                         </div>
                         <div className='ml-4 capitalize'>
                             <h6><strong>Graphic Design</strong></h6>
@@ -30,7 +30,7 @@ const HomepageSuggestiongroup = () => {
                     </a>
                     <a href='#' className='flex items-center mb-6'>
                         <div>
-                            <img src="https://images.prothomalo.com/prothomalo-bangla%2F2022-10%2Ff717609c-4123-4215-a5a5-104aca30db48%2FPakistan_1.jpg?auto=format%2Ccompress&format=webp&w=640&dpr=1.0" alt="" className='rounded w-16 h-16' />
+                            <img src="/user.jpg" alt="" className='rounded w-16 h-16' />
                         </div>
                         <div className='ml-4 capitalize'>
                             <h6><strong>Graphic Design</strong></h6>
