+++ src/pages/login.js
@@ -1,64 +1,125 @@
-import React from 'react'
-import 'bootstrap/dist/css/bootstrap.min.css'
-import Button from 'react-bootstrap/Button'
-import Form from 'react-bootstrap/Form'
-import Row from 'react-bootstrap/Row'
-import Col from 'react-bootstrap/Col'
+import ApplicationLogo from '@/components/ApplicationLogo'
+import AuthCard from '@/components/AuthCard'
+import AuthSessionStatus from '@/components/AuthSessionStatus'
+import Button from '@/components/Button'
+import GuestLayout from '@/components/Layouts/GuestLayout'
+import Input from '@/components/Input'
+import InputError from '@/components/InputError'
+import Label from '@/components/Label'
+import Link from 'next/link'
+import { useAuth } from '@/hooks/auth'
+import { useEffect, useState } from 'react'
+import { useRouter } from 'next/router'
 
-const login = () => {
+const Login = () => {
+    const router = useRouter()
+
+    const { login } = useAuth({
+        middleware: 'guest',
+        redirectIfAuthenticated: '/dashboard',
+    })
+
+    const [email, setEmail] = useState('')
+    const [password, setPassword] = useState('')
+    const [shouldRemember, setShouldRemember] = useState(false)
+    const [errors, setErrors] = useState([])
+    const [status, setStatus] = useState(null)
+
+    useEffect(() => {
+        if (router.query.reset?.length > 0 && errors.length === 0) {
+            setStatus(atob(router.query.reset))
+        } else {
+            setStatus(null)
+        }
+    })
+
+    const submitForm = async event => {
+        event.preventDefault()
+
+        login({ email, password, remember: shouldRemember, setErrors, setStatus })
+    }
+
     return (
-        <div className="container-fluid bg-white vh-100 flex items-center fluid={sm}">
-            <div className='Row flex justify-center'>
-                <Col md={8} sm={12}>
-                    <Row>
-                        <Col className='p-0'>
-                            <div className="card text-center">
-                                <div className="card-body shadow-2xl">
-                                    <Form>
-                                        <Row className='p-3'>
-                                            <h2 className='pb-5 font-bold'>Sign in</h2>
-                                            <Col xs={12}>
-                                                <Form.Group className="mb-3">
-                                                    <Form.Control
-                                                        type="email" className='p-2'
-                                                        placeholder="Email or Mobile"
-                                                    />
-                                                </Form.Group>
-                                            </Col>
-                                            <Col xs={12}>
-                                                <Form.Group className="mb-3">
-                                                    <Form.Control
-                                                        type="password" className='p-2'
-                                                        placeholder="Password"
-                                                    />
-                                                </Form.Group>
-                                            </Col>
-                                            <Form.Group className="mb-3 text-left" controlId="formBasicCheckbox">
-                                                <Form.Check type="checkbox" label="Check me out" />
-                                            </Form.Group>
-                                            <div className="d-grid gap-2">
-                                                <Button variant="primary" size="lg">
-                                                    SIGN IN
-                                                </Button>
-                                            </div>
-                                        </Row>
-                                    </Form>
-                                </div>
-                            </div>
-                        </Col>
-                        <Col className='p-0 flex items-center bg-emerald-700'>
-                            <div className='p-5 text-white'>
-                                <h4>We are more than just a company</h4>
-                                <p>
-                                    Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat
-                                </p>
-                            </div>
-                        </Col>
-                    </Row>
-                </Col>
-            </div>
-        </div>
+        <GuestLayout>
+            <AuthCard
+                logo={
+                    <Link href="/">
+                        <a>
+                            <ApplicationLogo className="w-20 h-20 fill-current text-gray-500" />
+                        </a>
+                    </Link>
+                }>
+
+                {/* Session Status */}
+                <AuthSessionStatus className="mb-4" status={status} />
+
+                <form onSubmit={submitForm}>
+                    {/* Email Address */}
+                    <div>
+                        <Label htmlFor="email">Email</Label>
+
+                        <Input
+                            id="email"
+                            type="email"
+                            value={email}
+                            className="block mt-1 w-full"
+                            onChange={event => setEmail(event.target.value)}
+                            required
+                            autoFocus
+                        />
+
+                        <InputError messages={errors.email} className="mt-2" />
+                    </div>
+
+                    {/* Password */}
+                    <div className="mt-4">
+                        <Label htmlFor="password">Password</Label>
+
+                        <Input
+                            id="password"
+                            type="password"
+                            value={password}
+                            className="block mt-1 w-full"
+                            onChange={event => setPassword(event.target.value)}
+                            required
+                            autoComplete="current-password"
+                        />
+
+                        <InputError messages={errors.password} className="mt-2" />
+                    </div>
+
+                    {/* Remember Me */}
+                    <div className="block mt-4">
+                        <label
+                            htmlFor="remember_me"
+                            className="inline-flex items-center">
+                            <input
+                                id="remember_me"
+                                type="checkbox"
+                                name="remember"
+                                className="rounded border-gray-300 text-indigo-600 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50"
+                                onChange={event => setShouldRemember(event.target.checked)}
+                            />
+
+                            <span className="ml-2 text-sm text-gray-600">
+                                Remember me
+                            </span>
+                        </label>
+                    </div>
+
+                    <div className="flex items-center justify-end mt-4">
+                        <Link href="/forgot-password">
+                            <a className="underline text-sm text-gray-600 hover:text-gray-900">
+                                Forgot your password?
+                            </a>
+                        </Link>
+
+                        <Button className="ml-3">Login</Button>
+                    </div>
+                </form>
+            </AuthCard>
+        </GuestLayout>
     )
 }
 
-export default login
+export default Login
