+++ src/styles/front.css
